A first try to build a pattern from a "real" text (extract from Conway's Game of Life wikipedia article). The credit goes to Prince who helped me to make the colored character list out using his favorite snake. We use python tu cut regular text slices also.I'm now trying to explain and summarize the commands we used in a file that may become a script.
We now have a problem to solve with the apostrophe and percentage characters. Here the backslash was not enough to escape.

The output postcards has strange colors of a cheesy spring. Perhaps a result of the color set, randomly generated. How to drive the association of characters and colors?

Soon on next postcard.